Rename OpenAPI File
