+++ src/commands/update.ts
@@ -1,8 +1,8 @@
 import { Command } from "commander";
 
-import { validateAndParseOpenApiSpec } from "../services/openapi-service";
 import { PluginService } from "../services/plugin";
 import { deployedUrl } from "../utils/deployed-url";
+import { validateAndParseOpenApiSpec } from "../utils/openapi";
 import { getSpecUrl, getHostname } from "../utils/url-utils";
 
 export const updateCommand = new Command()
